added some pizzaz and updated dash to work
