+++ client/src/components/Nav/index.js
@@ -20,7 +20,7 @@ const styles = {
     minWidth: '200px'
   },
   header:{
-    backgroundColor: '#F2A154'
+    backgroundColor: '#F7F6E7'
   }
 
 
@@ -34,7 +34,12 @@ function Nav() {
         <ul className=" d-flex flex-row justify-content-around fs-3 fw-bold" style={styles.width}>
           <li className="mx-1">
             <Link to="/awards-page">
-              Award List
+              Award List |
+            </Link>
+          </li>
+          <li className="mx-1">
+            <Link to="/dashboard">
+              Dashboard |
             </Link>
           </li>
           <li className="mx-1">
@@ -50,12 +55,12 @@ function Nav() {
         <ul className="d-flex flex-row justify-content-around fs-3 fw-bold" style={styles.width}>
           <li className="mx-1">
             <Link to="/signup">
-              Signup
+              Signup |
             </Link>
           </li>
           <li className="mx-1">
             <Link to="/login">
-              Login
+              Login |
             </Link>
           </li>
           <li className="mx-1">
@@ -72,8 +77,8 @@ function Nav() {
     <header className=" d-flex flex-row px-1 " style={styles.header}>
       <h1 style={styles.blueColor}>
         <Link to="/">
-          <span role="img" aria-label="shopping bag"></span>
-          Creative App Name
+          <span role="img" aria-label="shopping bag"><img src="..\images\badges\Flicks-on-the-Record.png"/> </span>
+          Flicks on the Record
         </Link>
       </h1>
